feat(TestHarness): changed TestHarness to Generic Class
Additionally, exposed some more data to public, including the once
regressed rng seeds.

For most, the only change you need to do is appening `<T>` of your
SimulationObject type. Hence I will not mark this as Breaking Change.

Part of the reason for this is the regression fix, and trying to expose
the final result. But all in all, I haven't been utilizing the
capability of reusing the TestHarness class for different
SimulationObjects